housing_prices/requirements.txt
numpy
pandas
xgboost
matplotlib
scikit-learn
scipy
pyyaml
easydict
graphviz
geopy